FIX: 채팅 페이지 tsx로만 변환
- 추가 작업 예정
